refactor logs column
